@@ -7,7 +7,7 @@ end
 
 
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem 'rails', '~> 5.0.2'
+gem 'rails', '~> 5.2.0'
 # Use Puma as the app server
 gem 'puma', '~> 3.0'
 # Use SCSS for stylesheets
@@ -18,6 +18,8 @@ gem 'uglifier', '>= 1.3.0'
 gem 'coffee-rails', '~> 4.2'
 # See https://github.com/rails/execjs#readme for more supported runtimes
 # gem 'therubyracer', platforms: :ruby
+
+gem 'bootsnap', '~>1.3.0'
 
 # Use jquery as the JavaScript library
 gem 'jquery-rails'
@@ -67,17 +69,23 @@ gem 'jquery-ui-rails'
 gem 'ckeditor'
 gem 'devise'
 gem 'openssl'
-gem 'haml',                    '4.0.7'
+gem 'haml',                    '~>5.0'
 gem 'will_paginate',           '3.1.5'
 gem 'bootstrap-will_paginate', '1.0.0'
 
 gem 'nokogiri'
+
+# HTTP requests
+gem 'httparty', '~> 0.13.3'
 
 # Application server
 group :unicorn do
   gem 'unicorn', '~> 5.1.0'
   gem 'unicorn-worker-killer', '~> 0.4.4'
 end
+
+# XLSX
+gem 'axlsx'
 
 # I18n
 gem 'gettext_i18n_rails'
